+++ src/components/resource-restart-modal.tsx
@@ -4,7 +4,7 @@ import { toast } from "react-toastify";
 
 import { camelToSnakeCase } from "../helpers/casing-helpers";
 import { useCurrentContext } from "../providers/current-context-provider";
-import { BaseModal, ModalButton } from "./modal";
+import { BaseModal, ModalButton } from "./base/modal";
 import { type ResourceBase } from "./resource-table";
 
 interface ResourceRestartModalProps<T> {
