+++ src/configurations/config-maps.tsx
@@ -1,7 +1,7 @@
 import type { V1ConfigMap } from "@kubernetes/client-node";
 
+import { TableCell } from "../components/base/table";
 import { ResourceTable } from "../components/resource-table";
-import { TableCell } from "../components/table";
 import { formatDateString } from "../helpers/date-helpers";
 
 export function ConfigMaps() {
