+++ src/networking/services.tsx
@@ -1,7 +1,7 @@
 import { type V1Service } from "@kubernetes/client-node";
 
+import { TableCell } from "../components/base/table";
 import { ResourceTable } from "../components/resource-table";
-import { TableCell } from "../components/table";
 
 export function Services() {
   return (
